Refactor image processing pipeline and improve modularity
- Replaced individual imports with a unified TensorFlow import (`import tensorflow as tf`) for better clarity and consistency across the codebase.
- Simplified `model.py` by refactoring the model-building logic to use `tf.keras.Sequential` for a more streamlined and modular approach.
- Enhanced `preprocess.py` by removing unnecessary functions and focusing on core preprocessing tasks:
  - Removed redundant functions such as `classify_item` and `detect_color` from `preprocess.py` and moved them to `process_and_insert.py` to maintain separation of concerns.
  - Refined image preprocessing by directly using `tf.keras.applications.mobilenet_v2.preprocess_input` for proper normalization.
  - Improved error handling in `preprocess_image` by adding a check for image loading failure.
- Updated `process_and_insert.py` to:
  - Handle classification and color detection within a single file for better encapsulation.
  - Include necessary imports (`KMeans`, `rgb_to_hsv`, etc.) and refactor the classification and color detection logic.
  - Introduce a dedicated function `process_and_insert_images` for handling the entire processing and database insertion workflow.
  - Retain the capability to insert processed data into MongoDB automatically.
  - Save preprocessed images to an output directory after processing.
- Improved overall code modularity and maintainability by ensuring that each file has a clear and distinct purpose.
